feat: ubah score ke integer

+++ src/main/java/com/lawencon/jobportal/model/response/ApplicationTrxResponse.java
@@ -3,7 +3,6 @@ package com.lawencon.jobportal.model.response;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Getter
@@ -13,5 +12,5 @@ public class ApplicationTrxResponse {
     private String stageName;
     private String statusApplicationName;
     private LocalDate date;
-    private BigDecimal score;
+    private Integer score;
 }
